% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h.sets.collection}
\alias{h.sets.collection}
\title{Hallmark Gene Sets from MSigDB at the Broad}
\format{A GeneSetCollection with 50 gene sets:
\describe{
 \item{names}{name of the gene set}
 \item{unique identifiers}{gene names for the genes in the gene sets}
 ...
}}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/index.jsp}
}
\usage{
h.sets.collection
}
\description{
A GeneSetCollection object created by using getGMT function from GSEABase. It is structured containing named gene sets and the genes the sets consist of.
}
\keyword{datasets}
