% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_logfc.R
\name{logFC}
\alias{logFC}
\title{Calculate metric for GSEA}
\usage{
logFC(cluster.ids, expr.mat)
}
\arguments{
\item{cluster.ids}{Named vector of the cluster identities assigned for each cell}

\item{expr.mat}{Expression matrix with rows as genes and columns as cells}
}
\value{
Returns the log fold change in gene expression for each gene for cells in a cluster vs cells not in the cluster
}
\description{
This function calculates the metric to be used for the gene set enrichment analysis. The most basic metric is the log fold-change in gene expression between a selected cluster and all other clusters, and is implemented here. Other possibile metrics could be used.
}
