% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw_gsea.R
\name{wmw.gsea}
\alias{wmw.gsea}
\title{Wilcoxon Mann Whitney Correlation Corrected GSEA}
\usage{
wmw.gsea(expr.mat, cluster.cells, log.fc.cluster, gene.sets)
}
\arguments{
\item{expr.mat}{Expression matrix for the dataset}

\item{cluster.cells}{List of clusters containing the assigned cells}

\item{log.fc.cluster}{Log fold change in gene expression between clusters}

\item{gene.sets}{Named list of gene sets containing the genes defining each set}
}
\value{
Returns the log fold change in gene expression for each gene for cells in a cluster vs cells not in the cluster
}
\description{
The function performs a correlation corrected GSEA for each cluster and returns a list of correlation corrected statistics and p values
}
