% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWMWgsa-package.R
\docType{package}
\name{scWMWgsa-package}
\alias{scWMWgsa}
\alias{scWMWgsa-package}
\title{scWMWgsa: What the Package Does (One Line, Title Case)}
\description{
This package takes single-cell gene expression data and lists of gene sets, and tests whether specific clusters have enrichment for the given gene sets compared to other clusters. To accomplish this, we implement a Wilcoxon Mann Whitney correlated-corrected statistic inspired by CAMERA to compare genes within the gene set versus genes outside of the gene set (see Camera: a competitive gene set test accounting for inter-gene correlation by Wu and Smyth, Nucleic Acids Research 2012 for more information).
}
\author{
\strong{Maintainer}: Anthony Cillo \email{arc85@pitt.edu}

}
\keyword{internal}
